requirements.txt
gradio==5.49.1
ultralytics
python-dotenv

torch
torchvision
torchaudio

opencv-python-headless
numpy
pillow
python-chess
cairosvg
fastapi
uvicorn[standard]
gunicorn